Support for microseconds order scheduling granularity
I'd like to send a feedback to SCHED_DEADLINE.
In our evaluation, SCHED_DEADLINE has not supported micro second order
granularity in its budget management. (The V6 also has same issue.)
The attached patch improves the scheduling granularity. But this patch
is created by the following steps.
(Not for development branch. Sorry for the inconvenience.)

 $ git clone git://github.com/jlelli/sched-deadline.git
 $ git checkout -b tmp remotes/origin/linux-stable-rt-dl

The base line for this patch is the following commit:
  -------------------------------------------------
  commit c7bd73156c6083f944c081703bf8f780af30b052
  Author: Juri Lelli <juri.lelli@gmail.com>
  Date:   Mon May 21 15:42:47 2012 +0200

    dl: double unlock balance inside find lock rq.
  -------------------------------------------------